requirements.txt
yaml>=0.2.5
pyyaml>=6.0.1
opencv-python~=4.9.0.80
numpy~=1.26.4
pytorch3d~=0.7.5
tqdm~=4.66.4